Remove hard-coded image and mask paths.
Move test images and masks into the .app and locate them as resources. This will make the app self-contained.
Add findfilesbyregexp() to utils to enable this.

Change view refreshes to be partial updates instead of refreshing the entire window contents.

Refactor the two view windows so that the animation timer code is in the common part and not the CG part.